refactor(skills): extract organization details state
